refactor(frontend): extract groups feature tabs into dedicated widget components
