+++ frontend/lib/screens/groups/active_tab.dart
@@ -1,6 +1,6 @@
-import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:syntrak/core/theme.dart';
+import 'package:syntrak/screens/groups/active_tab_widgets.dart';
 
 class ActiveTab extends StatefulWidget {
   const ActiveTab({super.key});
@@ -10,36 +10,13 @@ class ActiveTab extends StatefulWidget {
 }
 
 class _ActiveTabState extends State<ActiveTab> {
-  // Mock data for challenges
-  final List<_ChallengeData> _challenges = [
-    _ChallengeData(
-      id: '1',
-      title: 'January Vertical Challenge',
-      description: 'Accumulate 10,000m of vertical descent in January',
-      dateRange: 'Jan 1, 2026 to Jan 31, 2026',
-      badgeText: '10K',
-      badgeColor: const Color(0xFFE65100),
-      icon: Icons.terrain,
-    ),
-    _ChallengeData(
-      id: '2',
-      title: 'Winter Explorer Challenge',
-      description: 'Visit 5 different ski resorts this season',
-      dateRange: 'Dec 1, 2025 to Mar 31, 2026',
-      badgeText: '5',
-      badgeColor: const Color(0xFF1565C0),
-      icon: Icons.explore,
-    ),
-    _ChallengeData(
-      id: '3',
-      title: 'Speed Demon Challenge',
-      description: 'Record a run with max speed over 80 km/h',
-      dateRange: 'Jan 1, 2026 to Jan 31, 2026',
-      badgeText: '80',
-      badgeColor: const Color(0xFFC62828),
-      icon: Icons.speed,
-    ),
-  ];
+  late final List<GroupChallengeItem> _challenges;
+
+  @override
+  void initState() {
+    super.initState();
+    _challenges = mockGroupChallenges();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -53,17 +30,10 @@ class _ActiveTabState extends State<ActiveTab> {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            // Design Your Own Challenge Card
             _buildDesignChallengeCard(),
-
             const SizedBox(height: SyntrakSpacing.md),
-
-            // Available Challenges Section
             _buildChallengesSection(),
-
-            // Banner
             _buildBanner(),
-
             const SizedBox(height: SyntrakSpacing.xl),
           ],
         ),
@@ -77,7 +47,6 @@ class _ActiveTabState extends State<ActiveTab> {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          // Label
           Text(
             'SYNTRAK SUBSCRIPTION',
             style: SyntrakTypography.labelSmall.copyWith(
@@ -87,8 +56,6 @@ class _ActiveTabState extends State<ActiveTab> {
             ),
           ),
           const SizedBox(height: SyntrakSpacing.md),
-
-          // Title
           Text(
             'Design Your',
             style: SyntrakTypography.displaySmall.copyWith(
@@ -106,8 +73,6 @@ class _ActiveTabState extends State<ActiveTab> {
             ),
           ),
           const SizedBox(height: SyntrakSpacing.md),
-
-          // Description
           Text(
             'Rally your crew with a custom Group Challenge. Your game, your rules.',
             style: SyntrakTypography.bodyLarge.copyWith(
@@ -116,8 +81,6 @@ class _ActiveTabState extends State<ActiveTab> {
             ),
           ),
           const SizedBox(height: SyntrakSpacing.lg),
-
-          // CTA Button
           SizedBox(
             width: double.infinity,
             child: ElevatedButton(
@@ -158,7 +121,6 @@ class _ActiveTabState extends State<ActiveTab> {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          // Header
           Text(
             'Available challenges',
             style: SyntrakTypography.headlineSmall.copyWith(
@@ -166,20 +128,14 @@ class _ActiveTabState extends State<ActiveTab> {
             ),
           ),
           const SizedBox(height: SyntrakSpacing.md),
-
-          // Challenges list
           ...List.generate(
             _challenges.length,
-            (index) => _ChallengeCard(challenge: _challenges[index]),
+            (index) => ActiveGroupChallengeCard(challenge: _challenges[index]),
           ),
-
-          // See all link
           Align(
             alignment: Alignment.centerRight,
             child: TextButton(
-              onPressed: () {
-                // Navigate to all challenges
-              },
+              onPressed: () {},
               child: Text(
                 'See All Challenges',
                 style: SyntrakTypography.labelLarge.copyWith(
@@ -218,7 +174,6 @@ class _ActiveTabState extends State<ActiveTab> {
       ),
       child: Stack(
         children: [
-          // Background pattern
           Positioned(
             right: -30,
             bottom: -30,
@@ -228,15 +183,12 @@ class _ActiveTabState extends State<ActiveTab> {
               color: Colors.white.withOpacity(0.1),
             ),
           ),
-
-          // Content
           Padding(
             padding: const EdgeInsets.all(SyntrakSpacing.lg),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.start,
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                // Tag
                 Container(
                   padding: const EdgeInsets.symmetric(
                     horizontal: SyntrakSpacing.sm,
@@ -254,8 +206,6 @@ class _ActiveTabState extends State<ActiveTab> {
                     ),
                   ),
                 ),
-
-                // Title and subtitle
                 Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
@@ -279,8 +229,6 @@ class _ActiveTabState extends State<ActiveTab> {
               ],
             ),
           ),
-
-          // Mountain icons decoration
           Positioned(
             right: 16,
             top: 16,
@@ -304,260 +252,4 @@ class _ActiveTabState extends State<ActiveTab> {
       ),
     );
   }
-}
-
-// Data models
-class _ChallengeData {
-  final String id;
-  final String title;
-  final String description;
-  final String dateRange;
-  final String badgeText;
-  final Color badgeColor;
-  final IconData icon;
-
-  _ChallengeData({
-    required this.id,
-    required this.title,
-    required this.description,
-    required this.dateRange,
-    required this.badgeText,
-    required this.badgeColor,
-    required this.icon,
-  });
-}
-
-// Challenge Card Widget
-class _ChallengeCard extends StatelessWidget {
-  final _ChallengeData challenge;
-
-  const _ChallengeCard({required this.challenge});
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text('${challenge.title} details coming soon!'),
-            duration: const Duration(seconds: 1),
-          ),
-        );
-      },
-      child: Container(
-        padding: const EdgeInsets.symmetric(vertical: SyntrakSpacing.md),
-        child: Row(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            // Badge
-            _ChallengeBadge(
-              text: challenge.badgeText,
-              color: challenge.badgeColor,
-              icon: challenge.icon,
-            ),
-            const SizedBox(width: SyntrakSpacing.md),
-
-            // Content
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  // Title
-                  Text(
-                    challenge.title,
-                    style: SyntrakTypography.bodyLarge.copyWith(
-                      fontWeight: FontWeight.w600,
-                      color: SyntrakColors.textPrimary,
-                    ),
-                  ),
-                  const SizedBox(height: 4),
-
-                  // Description with icon
-                  Row(
-                    children: [
-                      Icon(
-                        challenge.icon,
-                        size: 14,
-                        color: SyntrakColors.textTertiary,
-                      ),
-                      const SizedBox(width: 6),
-                      Expanded(
-                        child: Text(
-                          challenge.description,
-                          style: SyntrakTypography.bodySmall.copyWith(
-                            color: SyntrakColors.textSecondary,
-                          ),
-                          maxLines: 2,
-                          overflow: TextOverflow.ellipsis,
-                        ),
-                      ),
-                    ],
-                  ),
-                  const SizedBox(height: 4),
-
-                  // Date range
-                  Text(
-                    challenge.dateRange,
-                    style: SyntrakTypography.labelSmall.copyWith(
-                      color: SyntrakColors.textTertiary,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// Challenge Badge Widget
-class _ChallengeBadge extends StatelessWidget {
-  final String text;
-  final Color color;
-  final IconData icon;
-
-  const _ChallengeBadge({
-    required this.text,
-    required this.color,
-    required this.icon,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 70,
-      height: 70,
-      decoration: BoxDecoration(
-        gradient: RadialGradient(
-          colors: [
-            color,
-            color.withOpacity(0.7),
-          ],
-        ),
-        borderRadius: BorderRadius.circular(12),
-        boxShadow: [
-          BoxShadow(
-            color: color.withOpacity(0.3),
-            blurRadius: 8,
-            offset: const Offset(0, 2),
-          ),
-        ],
-      ),
-      child: Stack(
-        alignment: Alignment.center,
-        children: [
-          // Starburst pattern
-          CustomPaint(
-            size: const Size(70, 70),
-            painter: _StarburstPainter(color: Colors.white.withOpacity(0.15)),
-          ),
-          // Content
-          Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Text(
-                text,
-                style: SyntrakTypography.headlineSmall.copyWith(
-                  color: Colors.white,
-                  fontWeight: FontWeight.w800,
-                  fontSize: 18,
-                ),
-              ),
-              Icon(
-                icon,
-                color: Colors.white.withOpacity(0.9),
-                size: 16,
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// Starburst pattern painter for badges
-class _StarburstPainter extends CustomPainter {
-  final Color color;
-
-  _StarburstPainter({required this.color});
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    final paint = Paint()
-      ..color = color
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = 1.5;
-
-    final center = Offset(size.width / 2, size.height / 2);
-    const points = 12;
-    final innerRadius = size.width * 0.25;
-    final outerRadius = size.width * 0.45;
-
-    final path = Path();
-    for (int i = 0; i < points * 2; i++) {
-      final radius = i.isEven ? outerRadius : innerRadius;
-      final angle = (i * 3.14159 / points) - 3.14159 / 2;
-      final x = center.dx + radius * math.cos(angle);
-      final y = center.dy + radius * math.sin(angle);
-
-      if (i == 0) {
-        path.moveTo(x, y);
-      } else {
-        path.lineTo(x, y);
-      }
-    }
-    path.close();
-    canvas.drawPath(path, paint);
-  }
-
-  double cos(double angle) => _cos(angle);
-  double sin(double angle) => _sin(angle);  double _cos(double angle) {
-    return (angle == 0)
-        ? 1
-        : (angle == 3.14159 / 2)
-            ? 0
-            : (angle == 3.14159)
-                ? -1
-                : (angle == 3 * 3.14159 / 2)
-                    ? 0
-                    : _cosInternal(angle);
-  }
-
-  double _sin(double angle) {
-    return (angle == 0)
-        ? 0
-        : (angle == 3.14159 / 2)
-            ? 1
-            : (angle == 3.14159)
-                ? 0
-                : (angle == 3 * 3.14159 / 2)
-                    ? -1
-                    : _sinInternal(angle);
-  }
-
-  double _cosInternal(double x) {
-    double result = 1.0;
-    double term = 1.0;
-    for (int n = 1; n <= 10; n++) {
-      term *= -x * x / ((2 * n - 1) * (2 * n));
-      result += term;
-    }
-    return result;
-  }
-
-  double _sinInternal(double x) {
-    double result = x;
-    double term = x;
-    for (int n = 1; n <= 10; n++) {
-      term *= -x * x / ((2 * n) * (2 * n + 1));
-      result += term;
-    }
-    return result;
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) => false;
 }
